Extracted Auth.Signatures into separate module
